UNITS - further Artillery energy revisions
More units having the Artillery energy mod applied to them.   This one covers more of the psuedo-artillery - that is, weapons that have artillery style ranges (for the most part) or extremely large AOE and Damage combinations.

Some of these units demonstrate an exception where certain T4 units provide a portion of the power  that would otherwise be demanded by the formula, for their weapons.