LUA: Random faction even spread by @joinr (#24)
joinr, AKA c04spoon#0623 on Discord, wanted an option for more even distribution of random factions at game start, and produced a patch file to make it happen; it can be found attached to the aforementioned issue. I modified it as follows:

- The option is now per-lobby instead of per-prefs, so it's transparent to all clients in a lobby.
- Added tooltips.
- Linted to follow code style conventions.

I also allowed the frontend lobby settings display to cut off overflowing option titles with an ellipsis specifically for the name of this new setting.