LUA - AI - base creation/timing issues
The mechanism used to pass new base information to an engineer is VERY time sensitive - and an engineer who is already onsite for a new base could get hung in a loop waiting for this information, simply because the thread which would feed him the data has already expired - he'd be waiting forever.

Fixed this, and addressed some other very lazy looping, so that not only do we repair this situation, but we eliminate a fair bit of slack in this operation.

Numerous logging options have been left behind - outcommented -- that were necessary to understand this complicated semaphore process and monitor it's sequencing.