LUA - AI - Air ratio triggers
A number of operations, like the construction of mobile AA, are triggered by Air Ratio.    The initial value for this was too low, with the result that the AI would often have little or even NO mobile AA, even though he was not in firm control of the air.

This value has been moved up, to trigger at a somewhat higher ratio - both for mobile AA production, the formation of pure AA platoons, and the desire for LOUD to want to build AA defenses.