BL & LUA: Fix damage buffs, make Superhot (damage node) use them
The Superhot (UEF T3 Damage Boost Node) from BrewLAN previously used the nonfunctional damageMod to work. In adapting it to use a damage buff, I discovered that damage buffs also don't work, since they called a method on Weapon that was never even declared (but doesn't throw any exceptions for some reason).

It's also worth noting that the damage buff is currently a 1.1 multiplier, and it stacks as such, which may not be desired, but it's not a disastrous situation balance-wise at the moment.