TM: Tone down Banshee sounds
The Banshee previously produced an enormous death explosion sound when it died and then a literal UEF nuke sound when it blew up. This made it audible across the map, and wrongly gave the impression that something more important was happening when one died.

Now its death produces the same death sound as the Pillar and its destruction produces that of the Titan and Percival. The original spirit is still there, but it's now in line with a T2.5 unit.