LUA: Add icon next to buildrate in unit views
What this textbox meant was to lost to me for months, and other users have had the same problem. I only found out once I looked at the code; now it should be somewhat clearer.

Not integrating this into statgroups is messy, but doing that leaves the layout even worse than it is in this implementation.

I also added drop shadow to the text itself for consistency and readability.