4D, LOUD, LUA, TM: Inter-mod excl., >1 excl. per combo
The unit exclusion functionality has been expanded such that a mod can now impart exclusions upon other mods.

Additionally,  one mod config setting can be tied to a mutually-exclusive set of exclusion blocks rather than just one, and an exclusion can be made inherent to a mod if the author doesn't want it to be configurable.