BLUEPRINTS.LUA - Reclaim reduction
This change is a general reduction in reclaim values - rather significantly - to about 60% of original.   
This is intended to reduce the penalization for being aggressive - and losing - thus indirectly making it more attractive versus turtling.    We'll watch this in play for a couple of months before we come to any great conclusion.   It does NOT affect natural reclaim - only wreckage.