LUA: Mod manager revamp
Work in progress. The most basic functionality is present, but dependencies/exclusivities, presets, "Disable All", and the LOUD Standard button are all out of order in this commit.

This new dialog creates a more compact (12 at once) foldered list of mods, displaying each in a line containing its name, icon, and a checkbox. Clicking on a line displays more detailed information about the mod on the right side of the dialog.

This will eventually have filters, but these aren't present yet either.