LUA - PriorityFunctions
Additional logging of priority changes and a revision to how PriorityFunctions are handled by the PlatoonFormManager.

Previously, PFM tasks did NOT get PriorityFunction reviews unless the PFM tasks were sorted.   This only occured when the primary status of the location changed (not very often).

Now - they are processed, inline with the PFM looking for builders to fill.