LUA - AI Engineer Defense Tasks
Because many of the defense builders are triggered to a higher priority by a PriorityFunction, we can slip them down to a priority where they'll have to compete with normal engineer production tasks (factory assist, etc).   This can translate into greater production of units and/or eco rather than defense building.

To dovetail with that thinking, we move the factory assist tasks up so that they come first now, prioritizing unit production over defense building - when ratios are low but we are not threatened