LUA: Prevent keybind list elements from overflowing down
Folding headers causes more headers to be shown in the dialog at once, causing the list to grow in height and spill over other elements.

The fix is a small reduction to the height of keyContainer, alongside a move of the close button upwards by 4px so the de-assign button fits the dialog more nicely.