LUA: Major fixes to de-assign key functionality
I hadn't given this button a full test; in doing so, I discovered that any clearing of a mapped key isn't reflected in-game until an application restart. Upon assignment, the new key will work fine, but the old key would too.

For now, that issue persists in the assignment button, but now the de-assign button does everything it's supposed to. It also checks if the selected action isn't already mapped, and Keymapper.ClearUserKeyMapping no longer has a useless action parameter.