LUA/LMISC: LOUD Debug Tools
A UI-only mod (under "Miscellaneous") which adds a menu for changing AI debug switches on the fly. It also enables toggling which intel threat types get shown when DisplayIntelPoints is on.

To make all of this work, AIBrain now has a function which allows modification of specific ScenarioInfo/loudutilities fields, exposed to the UI via a SimCallback.

Other changes of note are that DrawPlanNode had to be extracted out of AttackPlanMonitor into a global function and loudutilities.intelChecks was made global and given an extra bool field in its value tables.