BL/TEX: Enforce consistency on factory_naval exp. strat. icons
In talks with Balthazar, it was found that experimental strategic icons ending in factory_navy fell out of line with the factory_naval icons in the previous 3 techs. These are going to be rectified as we move to fill our lineup of strat. icons completely with Balthazar-made versions; this update takes the opportunity of this week's planned brewlan.scd update to normalize the Souiya.