LUA: AI chat multi reporting, better "give engineer" command
Any transmission to "All" in chat with one of the tokens "aicheat", "ailevel", or "aimult" anywhere will lead all AI to report their multipliers. This, as far as I can tell, is the fastest user flow for this use case without possibly breaking into the conversation of human players.

Additionally, if "engi" appears in a whisper to an AI, that will be read as an engineer request. The old command "give me an engineer" has been removed.