LUA: Make ratio/time ACTs completely cohesive
In vain attempts at optimization, these two functions had some not-very-sane interactions with each other and other AI logic:
- Ratio ACT would not factor into aiBrain.CheatValue.
- Time ACT would clobber ratio ACT's effects.

Now, brains store at least 1 and up to 3 cheat values: one for the baseline (which never gets changed), and one for each ACT the brain subscribes to (BaseCheat has been expunged). aiBrain:TotalCheat() adds all of these 3 which are present together, and this returned value is used in SetArmyPoolBuff, is printed when an AI is asked its mult, and goes into the SpawnWave too.

TotalCheat is not getting propagated to other consumers of the multiplier to reduce some overhead, avoid drastic difficulty spikes (i.e. new unit buff) and prevent unforeseen cascading effects on other AI behaviour.

To conclude, the new ACTs should now be completely predictable and transparent to the player.