LUA: Colour transfer between players, observers in lobby
Previously, trying to change from a player to an observer and back again led to a GUI collapse because of what looks like a negligent find & replace.

Giving the player a default or random colour so that it's not nil when they return to being a participant would have been easy, but it was not much more difficult to store the desired colour on a per-observer basis so that new joiners who start as observers but then convert to players can have their prefs-stored colour.