BL: (R&D) Make BrewRND a non-static member of AIBrain
This table, which encapsulates all an AI's research logic, was previously global to all AIBrains. I'm guessing there was intent behind this, since BrewLAN trunk does not do this, but it was an unreasonable intent without a full implementation. Making all AI share the same research state would have broken this mod in free-for-alls, and the way it worked before this commit caused every brain's initialization to reset the entire table.