BL/LUA: (R&D) Dynamic icon support for research items
In mainline BrewLAN R&D, what few research items get generated also get their own unit icons; a duplicate of the one depicting the researched unit.

This works fine if the only research items are for T1 and T2 pgens but not for an unknown number of units which we may opt to lock behind research. What's more is that Balthazar already suggested an R&D/Lucky Dip interaction, in which case we either need this commit or icons for every possible unit which might get restricted. The latter is unacceptable.

Thus, research items get Display.IconName assigned the ID of the unit they unlock (except for Tech Level researches). In select parts of the construction GUI, unit view, UVD rollover, and spawn menu, this ID gets checked and can be used to find the appropriate icon.

This occupies a few extra flops whenever this check has to be made (the amount of file IO is the same as before) and a few extra bytes for saving the strings in the BPs. Note that because avatars get their icons from the player faction, they still appear empty under certain edge cases. Maybe labs get their own avatar tab eventually.