BL: (R&D) Overhaul initial experimental restrictions (again)
Now these should be sensible regardless of which mods the user has enabled; each unit pack except Wyvern and 4D (being too small) has one experimental restriction for each faction. Generally these are still the heaviest of the heavy T4s. For BrewLAN it is the full range of T4 shields, and for vanilla it is each of the T4 artillery.