BL/LMISC: R&D/LuckyDip interaction
Balthazar suggested this idea weeks ago and I liked it enough to try implementing it. This is a minimum implementation; it isn't unoptimized, it has not been stress tested, I don't know how the AI will handle it, and while there is support for disabling R&D's unit restrictions so that only LDip's are applied, it's not in the config file yet. At this point, I just want this code committed now since this change is getting bloated enough as it is.