LUA - LAND Attack formations
With the work done on PARSEINTEL, a better understanding of threat interactions allows us to better tune a number of parameters for land platoons.
One new feature is the THREATMAXRATIO - this allows a platoon to take on greater threat when doing target assessment - and this in turn allows it to get better use of the other filters during target selection (Structure and Unit counts).   In general this results in a better and much wider choice of targets for the platoon.
A new, larger AA Attack platoon
Elimination of AA Attack Water - new platoon is universal