BLOPS: Polish Inferno display strings
The casing of the word "flamethrower" is well-defined and should be adhered to. "Tank" is a debatable naming choice; it has more health than a Pillar but can be perceived as being more fragile due to the effects of volatility in numbers. It still seems suitable for an assault-purposed self-propelled gun on tracks, however.