LUA/BL/LMISC: Overhaul unit icon framework
My impression of the old code which handed out unit icon paths to parts of the GUI which needed them is that it was founded on the limitations of a mod and integrated that way, but we don't have those limitations. There's also a distressing amount of code meant for unit build icons which don't exist in FA.

Blueprints.ModBlueprints() now seeks out icons from usermods and makes reference to them in unit blueprints. A one-stop function in GameCommon retrieves either that, or IconName (from R&D), or the icon in textures.scd, or a placeholder.

Additionally, legacy icon support for SupCom 1 in BrewLAN's construction.lua hook has been removed, since LOUD supports only FA.

Finally, the problem stopgapped by 3d4ced90 has been solved, so it's been partially reverted.