LUA - PARSEINTEL
Moved from an 8 stage process to a 15 stage process.   This allows me to separate the individual threattypes more so that I never have to process more than 3 in a cycle - and most times - only 1 - AIR.

This allows me to reduce the cycle time from 6.4 seconds down to 4.5, giving a much greater responsiveness to those types that need it - AIR especially.     Multiple threattypes wiithin a cycle are also more widely separated (5 ticks vs 3 ticks), better spreading any congested loads - and allowing the display routines to be more clearly delineated when the debug toosl turn them on.

The display time parameters have been better tuned to this cycle as have the TIMEOUT parameters.