LUA: Workaround for order button overflow
In a test game, I attempted to select a Custodian, some Tech 3 Engineers, and my double-drone-equipped ACU, and observed that the normal bottom panel didn't appear. This was caused by the presence of more possible orders than could fit in the order button grid (overcharge, Custodian jamming, 2 drone toggles, reclaim, capture, repair).

There aren't any nice ways to accommodate more orders than we already can, as this UI is cramped as it is on 800 * 600 displays. This is a dead-simple fix; some orders just get left behind.