ALL UNITS: Adopt new experimental icon naming scheme
'icon_experimental_land_directfire' is now 'icon_land4_directfire', for instance.

Also note: Kraptor, Mayhem, Avalanche, and Suedath are now directfire2. The Avalanche used to be 'mobilefactory' and the Mayhem used to be 'missile' but I opted to change them to indicate that they are among the heaviest in their tech level. I'll be testing to all of this to see if I can find obvious room for improvement in the readability department.