LUA - AIBehavior - Bombers/Gunships
Bomber and gunship target searches are using their AntiSurface threat to judge if it's safe to go after targets - which is wrong, but has always been done since they don't really have any AntiAir of their own.

We'll now only use 1/3 of it for bombers - and 1/2 for gunships.    This should make target selection somewhat more judicious when enemy anti-air assets in the area are high.