LUA - Cheat adjustments
In keeping with a long standing goal of reducing cheating requirements, we're reducing the additional buffs to HP, regen and such from the Cheating modifier.

Originally at par with the cheat value, that buff has been steadily reduced over time, to that point that we are today reducing it to 33% of the cheat value.

Of course, sorry to say, but we are enacting these buffs into the ACT system to provide conformity with our basic AI cheats.

The economic consumption bonuses, which were at 75% of the cheat value, have also been reduced - to 67%.  

The final piece of this update is a shift in the T1/T2 HP adjustment, applied to mobile units.   The T1 bonus is dropping from 27.5% above blueprint value, to just 22%.   This represents our assessment of the improved overall balance of T1 versus T2, and it's position in the game.

Removed some log tags.