LUA - AI Engineers - Mass building
One of the weak points of mex building for engineers is that they'll often select the same goal for building an extractor resulting in a 'conga line' behavior which is wasteful.  

To combat this, the process will now return a list of possible choices, and one will be randomly selected.   This value is now part of the Builder Data for the engineers.

Numerous renaming of engineer buildernames to better identify range and looping choices

Re-prioritized many engineer builders.

Added looping for Hydrocarbon building.