LUA - AI Distress Response
This is a complete rework of the Distress Response system to provide greatly improved responsiveness and logging for analysis.

A BIG fix for retrieving the threat of your own platoons.   Unlike other threat systems, the function for getting threat on your own platoons requires a DIFFERENT handle to retrieve - ie.  rather than AntiSurface, which is widely used all over, this function requires Surface, and in fact, only responds to the threats by actual name - so Surface, Air, Economy, Sub - just as in the Defense system of blueprints.