LUA - AI - Primary Attack Base
Found bug where the primary base could be declared but not actually exist.    This caused a number of platoons to be paralyzed in reinforcement roles.

To fix, each time the primary base check is run - any existing primary is cleared - and additional check in place to assure that it's an active position.