LUA - AI Attack Planner, Primary Base Logic, Pathfinder tweaks
The AI Attack Planner is responsible for setting some primary goals for the AI - therefore it needs a robust logic for finding points suitable for attacking the goal - this involves a lot of pathfinding, and a lot of logic to cover failure situations.    The results of this logic will impact the PRIMARYBASE logic.

The PRIMARYBASE logic tells LOUD which of his bases is best suited for attacking the primary goal - and he moves his reinforcements accordingly - which is why the ATTACKPLANNER logic is so crucial.

Fixed issue where PRIMARYBASE logic would fail if a base was declared PRIMARY but then quickly destroyed.   Addressed issue where if the ATTACKPLANNER was unable to create a viable path to the goal, the PRIMARYBASE logic would fail - it will now declare whichever position is physically closest (as opposed to path distance closest when it works properly).

A lot of pathfinder calls have been split up - to insure there is no overlap between pathfinder calls.   Due to the latency in returning results, the ATTACKPLANNER would sometimes get path results for a different stage of the process.

PATHFINDING has been adapted to return more accurate path length values - it was previously ignoring the path distance from the start position to the first node - and the path distance from the last node to the final destination.   These values are now part of the total pathing distance.