LUA - MovePlatoon function
This function gets some expansion so that Support and Guard squads have better functionality within the platoon as it moves.    This should curtail behaviors like the shields running ahead of the pack.

This change necessitates the revision of some of the platoons and the squads within them.  

Since MovePlatoon only applies these upgraded behaviors to aggressive platoons - some platoon builders have had the AggressiveMove flag turned on.