UNITS - Base - Skirt Adjustments
Adjusted skirts of Aeon & Cybran Naval factories to be inline with skirts of other structures/units.

Adjusted skirts of Aeon, Cybran & Seraphim Naval factories so that the back of the factory is next to the skirt allowing the rest of the skirt area to be used as the build area.

Adjusted skirts of UEF Atlantis & Seraphim Ythotha to be inline with skirts of other structures/units & reduced the skirt size to center them as close as possible.