LUA - Gunship formations, behavior
Tightening the Gunship leashes a little bit - and moving the difficulty down so that gunships will stay a bit closer to home and not drop themselves into bad situations as often.
I've also altered the spread of secondary targets to focus more upon shields and AA at the primary target area - and narrowed the search radius slightly.