LUA - Base counting
Long ago, we used to have limits on how many land or naval bases we allowed - as a means of keeping performance in check - that is no longer required - and in fact - keeps us from distributing spare engineers and SACU to other bases - since we were only counting production bases.   
Now - ALL types of bases, production or defensive points, are counted.