LAU - Energy Weapon charge cycle
Thanks to Phoenix, we discovered a flaw where a unit with an energy charge cycle - could bypass that by moving the target during the firing cycle - allowing almost instant restart of the firing cycle - without need for a recharge.
The solution - move the start of the energy charge cycle to a point where the unit begins to fire, rather than waiting for the salvo to discharge.
This MAY cause some issues down the road, where we were counting on some other delays to impact the firing cycle - as now the charge cycle may shorten the firing cycle for some of those that were - we'll have to keep an eye out for those.
The upside to this change is that it should spread the recharge out over the ENTIRE firing cycle - thus reducing the spot nature of the energy drain.