LUA - Engineer Platoons reduced
Removed several different engineer platoons and boiled it down to just two (and only because I didn't want to go to some effort).

Likewise - paired down ACU platoons

All jobs that were associated with the older platoons now have their AI plans assigned directly as part of the platoon definition.   This makes it MUCH easier to locate which behavior is assigned to which job.