LUA - RETREAT AI
Previously, retreat behavior was built into the various platoon behaviors - many had it, but others did not.   Also, retreat could only be triggered at various points in the platoons behavior cycle.

We now create a 'behavioral' function - that runs on it's own, and can be introduced for any platoon.  Since it's on it's own - it can trigger a retreat behavior at any time in the life of a platoon, and it allows us to remove any code surrounding that process from the platoon behaviors themselves.

Added 'RetreatAI' to numerous platoons