LUA - GuardPoint improvements
A real cleanup and revisal of the core behavior - taking advantage of RetreatAI, category based targeting,  data reduction,  and PlatoonFindTarget function

Removed retreat behavior.

Like AIScouting functions, any scouts will now form a growing scouting pattern around the point, based upon the number of scouts in the formation.

The guardRadius of the platoon itself will grown and shrink with the size of the formation.