LUA - EngineerBuildMassAdjacencyAI
Revised function to accept a category to test for (ie. MASSSTORGE, ENERGYPRODUCTION) so that the function will be able to understand building different types of adjacency - previously only MASSSTORAGE.

Added AdacencyStructure value to the Construction table of Engineers so that different units can be counted as part of the point selection requirements.