LUA - ThreatMax on Engineers
ThreatMax is used to keep engineers from doing stupid things - BUT - the default threat level is 10 - and that's really small.   The real downside is that it can take quite a while for even modest threat to decline down to 10.

Putting specific threatmax values on tasks allows some tasks to be do-able before that point, and this better helps the separation between engineers as well, since T1 engineers are, and should be, more sensitive to threat than a SACU.

Implemented threatmax values amongst most of the defense building tasks.

Raised the default from 10 to 20.

Improved a failure message to give UnitID details.