LUA - Air Refit logic
Some timing adjustments and a restructure of the Callbacks that trigger it - Air units were often being inserted into TWO or more refit platoons.

Callback now checks for an InRefit semaphore the prevents this - a warning will be spewed to the log if the code tries to do this - for further examination.