LUA - BRAINCONDITIONMONITOR timing
Like PARSEINTEL, the BCM (BrainConditionMonitor) is a timing critical process - and I wanted it to be accurate in terms of it's length - this has importance to those functions which rely upon checked conditions.

I also wanted to introduce a little more 'slack' when there are more players involved (AI or Human).