LUA - BUILDERMANAGER timing
Once the timing of the BCM is understood, we can better synchronize the BuilderManager (which governs the Platoon Form Manager (PFM)).

The importance of this is we insure that all build conditions that are not 'instant' have been checked and are up-to-date in a certain amount of time.    This allows us to run the PFM every bit as hot as the BCM, insuring the most expeditious creation of platoons.

With this change - the PFM can now cycle twice as fast as previously, for all types of bases, resulting in much greater responsiveness overall = less base clutter and general buildup = more units in the field at any given time.