LUA - AI Base selection for maps 20k or less
There are just some maps where the STD base layout is too big.   Often these maps are 20k or less, and have some water component, or very congested terrain.  
In the case of the map Adaptive Hutts Clutch - both were true - and the AI could not place a Quantum Gate using the STD base layout.  

Solution - I've altered the base selection process to detect the water component better - and lean towards the smaller base on any 20k with water on it.