LUA - Performance work
After some discussion with JIP about how LUA utilizes lookups (. syntax in variables), I went and expounded on it in some key areas.   
I already do a great deal of this kind of optimization, but I took the opportunity to take these specific places to another level of this kind of work.

Along the way there were also a few more places where TABLE commands were not localled but could be.