LUA - even more
Aside from continuing to rework the usage of LOUDGETN, there is also some minor data reductions going on, in this case essentially removal of brain values that have no continued purpose and are just consuming memory if sitting around.  If required, the value will be recreated, and removed when finished.