LUA - Sera Air Factory Engineers dying
Engineer built from Sera Air factories were not moving off the pad - and would die when the next unit was started (pushed them off the edge).

This was caused by clearing the UnitBeingBuilt field on the factory when the build was complete.   The Sera Air Factory was using custom code for after the build, to detach the engineer and pass it off to a custom roll-off event.

This custom code has been moved into the default units, and the UnitBeingBuilt field is now only cleared when the StructureUnit returns to an Idle State.