LUA - Starting mass/energy bonus on ACU
When cheating and/or being outnumbered, the AI ACU will receive an energy and mass bonus equal to the ratio of the cheat and the outnumbered ratio.

This fix insures that the bonus applies properly, as previously, it was being given at full value even when outnumbered.

The length of the energy bonus has been extended to 45 seconds, after which it is removed - and the mass storage bonus is enacted.