LUA - Initial Mass point dispersion
Up until now, the AI has been hit and miss with respect to allocating mass points in a fair fashion between Allied AI.   

Now, each AI will create a list of mass points that it will focus upon obtaining first, prior to using the original logic.   This list is created in a common thread, so there is no overlap, and each AI picks those closest to it - provided it has not already been taken.    This results in a much better dispersion between Allied AI with regards to starting mass.

Because the setup for this process occurs very early on, we had the opportunity to clean up the ScenarioInfo data a bit - since we utilize this data to prepare the mass point selection tables for each 'team'.

Revised some utility functions with revised data

The AIBuildStructure module now employs the initial mass point starting list for all mass point lookups at the start of the game, until the list is exhausted.