LUA - Amphibious Hover balancing
Until now, we never assigned any particular values to the hover units and what being a hover means to the unit.   Now, we introduce a basic cost difference of 10%+ energy.   This is a perfunctory cost applied to ALL hover units, including engineers that hover.  In of itself, it shouldn't be much of an shift.   The real balance shift comes in it's relation to naval units.

Naval units, and their production, suffer a number of penalties, partly due to their mass heavy nature, and from the lack of any kind of mass or energy adjacency for naval factories.    This is an unfavorable relationship for the entire class - and today, we'll introduced a very mild shift in that, a global 10% reduction in the energy costs of all NAVAL unit production.

For now - we'll effect these changes thru blueprint.lua adjustments, and we'll start the process of editing the many unit blueprints, thru all the assocaiated mods, as we find the time.