LUA - Construction of Artillery
Like all STRUCTURES, Artillery and TML are placed into the Structure Pool when completed.   This is a bit of a problem, and we encountered it before with NUKES - the PlatoonFormManager simply cant find them to put them into a platoon to run their behavior.

The solution was already present for NUKES - we detect them at complete and we DONT move them into the Structure Pool, we leave them in the Army Pool, where the PFM can find them.

For now - we'll do this to STRATEGIC ARTILLERY and we'll go back and add TML later.