LUA - AIAttackUtilities - AIFindTarget...
The AIFindTargetInRangeInCategoryWithThreatFromPosition() function (wow - what a mouthful) has been updated in a number of ways.

If no threat data is provided - the routine will simply report the first valid target it finds closest to the position.   
Also - the function now correctly uses the squad provided to the function, to validate targets.   This is important, since the Artillery Director will call this function with the 'Artillery' squad, unlike all the other calls which use the 'Attack' squad.