LUA - Base mode switch
Bases can operate in Amphibious or Land mode, and this generally controls production.    This condition is controlled by the path to the current goal.

We've expanded how this condition affects not only Production of units, but his behavior locally in terms of which formations will appear.   The idea of this change is to make him much more sensitive to local enemy threats - and to have much more focus on dealing with them.

The change in mode will now affect ALL production bases, allowing amphibious and the cheaper land units to be produced as required - instead of always producing exclusively land or amphibious units.