LUA - Threat based Build Conditions
One of the primary triggers for defensive behavior is the presence of local threat factors.   A lot of recent data has allowed me to better analyze this, not only for which amount, but distance - of land based threats.

Previously, we were adjusting these threat evaluations by the LAND RATIO and the cheat multiplier (at game start) - as a result - if cheats were high, or ratios got somewhat high, LOUD became overly tolerant of local threats.

Threat distance is now only modified by the base cheat level - which makes him sensitive to threats that are even further away.   We no longer are concerned with the ratio - if it's close - LOUD will go after it.

This better data is now reflected in the conditions I use