LUA - Mass Fab Thread
The Mass Fab Thread controls the on/off state of the AI's mass fabs in accordance with the economy.   However, while observing them, I noticed that the upgraded versions were NOT being controlled by the thread.   I tracked the issue to the self-upgrade process.

In the end, I decided to make the thread a class object instead of relying upon the construction process, so that it will start up on ALL fabricators whenever and however they may be built - and it will identify if they are AI controlled or not.

- removed original MassFabThread from AIBehaviors
- MassFabThread now built into - and launched by the MassFabricationUnit class, upon unit creation, as part of DefaultUnits