LUA - Application of 'Outnumbered' ACU buff
The outnumbered buff is applied to the ACU at the start of a game, awarding bonus E storage, and the E to fill it, relative to how outnumbered the AI is.    This bonus lasts 45 seconds and is then replaced by an equivalent mass storage bonus - which is then also removed.   The purpose of this buff is to keep a badly outnumbered AI competitive thru the first few moments of the game.

However, when there is no outnumbered situation, this buff was still trying to activate (albeit with no bonus) - but - when the changeover to normal rates occur, the storage is refilled.  

This process is now properly bypassed when no outnumbered situation exists.