LUA - AI Energy Builders
T1 Power gen builders shifted into a position to be the 'fourth' builder likely to be selected at game start - when power is critical.    This is especially true in outnumbered situations where the 4th engineer is likely to be available.

Also, when assisting Energy builds, we'll shorten the assist time from 120 down to 100.
Energy Reclaiming starts at an earlier trigger point.

In normal circumstances, the ACU fills this role, but only after some initial MEX builds - it an outnumbered situation, we want to put more emphasis on avoiding any kind of power stall.