LUA - Engineer pathfinding re: plateaus
One of the complaints about engineers and plateaus revealed a logic error with engineer pathfinding.   This was resulting not only in extra long paths - but - in erroneously thinking that isolated extractors on plateaus could actually be reached.

The engineer will now correctly find the short route when path seeking, and will correctly recognize that such extractors are not reachable - and that transport should be called.

Installed some active logging for now to further report on these kind of conditions.