LUA - Loiter Point for AIR behaviors
This is the introduction of loiter points for AIR formations.   The idea of loiter points is simple - get air formations out into the field - rather than hanging around the base.

This increases the likelihood of getting targets, and responding to pertinent threats quickly, being much closer to the kind of activity that those formations will be looking for.

Each type of formation, ASF, Bomber, Gunship and Torp Bombers, has a unique function tuned for it's ideal operating conditions.