LUA - Builder Manager process
A resequencing of BuilderManager operations in order to better react to PriorityFunction changes.  Such priority changes will now be processed ahead of evaluating a task, allowing them to react immediately, rather than waiting for the next iteration of the Manager.  Tasks that drop to priority 0 will be removed from the table, and the table rebuilt before the next iteration of the manager, eliminating the table.remove that existed in the process, and allowing the processing of all tasks in the one pass.

Some data reductions as well, and numerous logging changes.