LUA - introduction of Artillery Shield to the AI
This work will allow the AI to build Artillery Shield (or equivalent) structures in their base layouts.

A good part of this work is utilizing ANTIARTILLERY category as a means of differentiating normal SHIELD units from these units.

Along with this introduction, a review of the eco triggers under which ALL the shields at a base are produced has been done.

A new builder for UEF, Cybran and Aeon has been introduced that will construct 3 AA Shieds in the MAIN base has been added.