LUA - Dummy unit merge
A fix for the bug I introduced when I added the ANTIARTILLERY category.   Normally, this would take place in the UNITS file, but since I don't want to patch that just for that tiny change, I'll introduce a merge BP here in the LUA.   Once UNITS is patched and released, I'll remove this file.