LUA - AI Submarine production
One of the underlying contributions to the buildup of large naval forces sitting around at a base is the lack of enough submarines to fill out the formations.   This will stimulate additional production.

To supplement this decision, the submarine requirement on naval formations has been lifted for the time being.