BO - Dreadnaughts and Bombardment ships
All the above classes have had the energy formula applied to their bombardment weapons.  NOTE:  This really is going to test those fans of these units - the power consumption is simply HUGE - and it should encourage a discussion about how it can be adjusted to make it more affordable to operate.  (ie - AOE, damage or range reductions).