LUA - IMAP rings
The IMAP ring system needs to be variable with map size, in order for those threat usages which require rings, to be consistent across map sizes.   There are quite a few places where 'rings' are used - we call them IMAPBlocks internally now, and we'll store the value, on the ScenarioInfo, when the game session is launched (when ParseIntel begins).