PROJECTILES - Big revisions for Projectile system
The Projectile mechanics are carrying around a great deal of extraneous information, and this provided an opportunity to reduce the data components, relying upon defaults where they exist.

In creation, a large number of tables were being created, for impact effects to be filled in.   Rather than do this, we'll only created the needed structure when we actually impact something - likewise - a lot of the data structures were carrying around blank tables and default scaling values (1).   Those have been culled where possible - reducing the data set size.

Just as importantly - the DamageData table was being created, and in quite a few cases, not being used - so we'll create this table ONLY when there is damage information - and we'll handle it entirely in the PassDamageData process.

One new process we've discovered in this work - the PassData function.  This one has escaped notice until now - and it's purpose is VERY edge case.   It's intended to carry data forward for some special projectile behaviors - such as the UEF projectile that creates a vision entity - OR - just as important - shells which have DoTDamage.   This is how those shells work.   The big discovery ?  Many shells were creating this table, without actually doing anything - in most cases, it was just a partial copy of the DamageData table.